requirements.txt
python-telegram-bot # for bot objects and handlers
openpyxl 
python-dotenv # for .env loading
fastapi # local webhook HTTP server
uvicorn # to run the ASGI server